Add files via upload
This is a simple and clean Login and Registration landing page built using HTML and CSS. It features a responsive layout, user-friendly form design, and a modern look with SVG icons. The page is split into two sections: one for user login and the other for new user registration. This project is part of my Web Development Journey